# VMS-6087 windows compile fixes
<==== Mercurial data ====>
Branch: vms_3.1_dev

+++ mediaserver_core/src/rest/handlers/merge_systems_rest_handler.cpp
@@ -350,7 +350,7 @@ int QnMergeSystemsRestHandler::execute(
     }
 
     nx::vms::discovery::Manager::ModuleData module(
-        remoteModuleInformation, {url.host(), remoteModuleInformation.port});
+        remoteModuleInformation, {url.host(), (uint16_t) remoteModuleInformation.port});
     owner->commonModule()->moduleDiscoveryManager()->checkEndpoint(module.endpoint, module.id);
 
     /* Connect to server if it is compatible */
